refactor: Move <TextElement/> to @chainlit/components
